backend/requirements.txt
fastapi[all]
uvicorn
psycopg2-binary
python-dotenv
